Moves brightcove player tests
