+++ packages/brightcove-video/__tests__/.eslintrc.json
