+++ packages/brightcove-video/__tests__/brightcove-player.web.test.js
@@ -2,7 +2,7 @@ import { View } from "react-native";
 import React from "react";
 import ReactDOM from "react-dom";
 import renderer from "react-test-renderer";
-import BrightcoveVideo from "../brightcove-player.web";
+import BrightcoveVideo from "../src/brightcove-player.web";
 
 describe("brightcove-video web component", () => {
   afterEach(() => {
