+++ packages/brightcove-video/package.json
@@ -15,14 +15,14 @@
   },
   "jest": {
     "preset": "react-native",
-    "rootDir": "../../",
+    "rootDir": "../..",
     "coverageDirectory": "<rootDir>/packages/brightcove-video/coverage/",
     "testEnvironment": "jsdom",
     "transformIgnorePatterns": [
       "node_modules/(?!(react-native|@times-components)/)"
     ],
     "testMatch": [
-      "<rootDir>/packages/brightcove-video/src/__tests__/*.test.js"
+      "<rootDir>/packages/brightcove-video/__tests__/*.test.js"
     ]
   },
   "repository": {
